requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
pkg-resources==0.0.0
pytz==2018.4
